requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
schedule
pandas==2.2.3
numpy<2
requests
python-dotenv
plotly==6.0.1
scipy
openpyxl
yfinance
